% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forage_production.R
\name{calc_forage_prod}
\alias{calc_forage_prod}
\title{Calculate Forage Production}
\usage{
calc_forage_prod(sample_wt, sample_units, sample_area, output_units)
}
\arguments{
\item{sample_wt}{The weight of the forage sample.}

\item{sample_units}{The units of the forage sample weight.}

\item{sample_area}{The area of the sampling frame. Most likely calculated using \code{\link{calc_plot_area}}.}

\item{output_units}{The units to return the forage production in. Either 'kg/ha' or 'lb/acre'.}
}
\value{
The forage production in the units provided, either 'kg/ha' or 'lb/acre'.
}
\description{
Calculate Forage Production
}
\examples{
sample_area <- calc_plot_area(dim_x = 50, dim_y = 50, units = "cm")
calc_forage_prod(sample_wt = 100, sample_units = "g", sample_area = sample_area, output_units = "kg/ha")

}
