% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forage_production.R
\name{calc_plot_area}
\alias{calc_plot_area}
\title{Calculate the Area of a Sampling Frame}
\usage{
calc_plot_area(dim_x, dim_y, units)
}
\arguments{
\item{dim_x}{The length of the quadrat in the x-dimension}

\item{dim_y}{The length of the quadrat in the y-dimension}

\item{units}{The units of the quadrat dimensions}
}
\value{
The area of the sampling frame in the units provided.
}
\description{
Calculate the Area of a Sampling Frame
}
\examples{
calc_plot_area(dim_x = 30, dim_y = 30, units = "cm")

}
