% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_utilities.R
\name{compare_file_hashes}
\alias{compare_file_hashes}
\title{compare_file_hashes}
\usage{
compare_file_hashes(old_file, new_file)
}
\arguments{
\item{old_file}{Path to the old file.}

\item{new_file}{Path to the new file.}
}
\value{
A Boolean value (TRUE/FALSE) indicating whether the file hashes match.
}
\description{
compare_file_hashes
}
\examples{
old_file <- tempfile(fileext = ".txt")
new_file <- tempfile(fileext = ".txt")
writeLines("Hello, world!", old_file)
writeLines("Hello, world!", new_file)
compare_file_hashes(old_file, new_file) # returns TRUE

old_file <- tempfile(fileext = ".txt")
new_file <- tempfile(fileext = ".txt")
writeLines("Hello, world!", old_file)
writeLines("Goodbye, world!", new_file)
compare_file_hashes(old_file, new_file) # returns FALSE

}
