% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/column_specs.R
\name{create_greenfeed_col_spec}
\alias{create_greenfeed_col_spec}
\title{Create a \code{readr} column specification for GreenFeed data.}
\usage{
create_greenfeed_col_spec(type = "preliminary")
}
\arguments{
\item{type}{Specifies the type of the greenfeed data to create a
column specification for. Options are \code{preliminary} and \code{verified}.}
}
\value{
A \code{readr} column specification for the preliminary greenfeed data.
}
\description{
\code{readr} column specifications are used to specify the data types of each
column in a data frame when reading in data from a file. This function
creates a column specification for GreenFeed data.
}
\details{
Column specifications are useful for ensuring that data is read in
correctly and that columns are parsed according to the correct data type.
For example, a column that contains dates should be parsed as a date
when reading in the data.
}
\examples{
\dontrun{
create_col_spec()
}
}
