% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_validation.R
\name{eid_prefix}
\alias{eid_prefix}
\title{Get the Prefix of an EID}
\usage{
eid_prefix(eid)
}
\arguments{
\item{eid}{A character string representing
the electronic identification number (EID) of an animal.}
}
\value{
A character string with the first three digits of the EID.
}
\description{
Get the Prefix of an EID
}
\examples{
eid_prefix("982012345678901") # returns "982"
eid_prefix("985012345678901") # returns "985"

}
