% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forage_production.R
\name{generate_forage_data}
\alias{generate_forage_data}
\title{Generate a Random Sample of Forage Production Data}
\usage{
generate_forage_data(n_rows)
}
\arguments{
\item{n}{The number of samples (rows) to generate.}
}
\value{
A data frame with random forage production data.
}
\description{
Generate a Random Sample of Forage Production Data
}
\examples{
generate_forage_data(10)

}
