% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate-data.R
\name{generate_greenfeed_data}
\alias{generate_greenfeed_data}
\title{Generate synthetic GreenFeed data for testing or simulation.}
\usage{
generate_greenfeed_data(n_rows, type = "preliminary")
}
\arguments{
\item{n_rows}{The number of rows of synthetic data to generate.}

\item{type}{The type of greenfeed data to generate.
The default is "preliminary". The other option is "verified".}
}
\value{
A CSV file containing synthetic GreenFeed data.
The data includes the following columns:
\itemize{
\item FeederID: A numeric value corresponding to the feeder ID.
\item AnimalName: A character value corresponding to the animal name.
\item RFID: A character value corresponding to the RFID.
\item StartTime: A datetime value corresponding to the start time.
\item EndTime: A datetime value corresponding to the end time.
\item GoodDataDuration: A time value corresponding to the length
of the observation in HH:MM:SS.
\item CO2GramsPerDay: A numeric value corresponding to the carbon dioxide
in grams per day.
\item CH4GramsPerDay: A numeric value corresponding to the methane
in grams per day.
\item O2GramsPerDay: A numeric value corresponding to the oxygen
in grams per day.
\item H2GramsPerDay: A numeric value corresponding to the hydrogen
in grams per day.
\item H2SGramsPerDay: A numeric value corresponding to the hydrogen sulfide
in grams per day.
\item AirflowLitersPerSec: A numeric value corresponding to airflow in
liters per second.
\item AirflowCf: A numeric value corresponding to correction factor
used to adjust gas measurements.
\item WindSpeedMetersPerSec: A numeric value corresponding to the wind speed
in meters per second.
\item WindDirDeg: A numeric value corresponding to the wind direction in degrees.
\item WindCf: A numeric value corresponding to the wind in cubic feet.
\item WasInterrupted: A logical value indicating whether the data was
interrupted by another animal. If the data was interrupted,
the value is TRUE; otherwise, it is FALSE.
\item InterruptingTags: A character value corresponding to the interrupting
RFID tags from the animals that interrupted the observation.
\item TempPipeDegreesCelsius: A numeric value corresponding to the
temperature of the pipe in degrees Celsius.
\item IsPreliminary: A logical value indicating whether the data is preliminary.
If the data is preliminary, the value is TRUE; otherwise, it is FALSE.
\item RunTime: A datetime value corresponding to when the observation
was processed by C-Lock's pre-processor
}
}
\description{
Generate synthetic GreenFeed data for testing or simulation.
}
\examples{
\dontrun{
generate_greenfeed_data(n_rows = 1000, type = "preliminary")
}
}
