% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_utilities.R
\name{generate_md5_hash}
\alias{generate_md5_hash}
\title{Generate md5 hash for a file}
\usage{
generate_md5_hash(file_path)
}
\arguments{
\item{file_path}{A path to a file.}
}
\value{
A md5 hash for the file.
}
\description{
Generate md5 hash for a file
}
\examples{
temp_file <- tempfile(fileext = ".txt")
writeLines("Hello, world!", temp_file)
md5_hash <- generate_md5_hash(temp_file)

}
