% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate-synthetic-data.R
\name{generate_rfid}
\alias{generate_rfid}
\title{Title}
\usage{
generate_rfid(seed = NULL)
}
\arguments{
\item{seed}{An optional seed for reproducibility.}
}
\value{
A numeric value corresponding to the RFID. The first 9 digits are all zeros. The last 15 digits are randomly generated.
}
\description{
Title
}
\examples{
generate_rfid(seed = 123)

}
