% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate-synthetic-data.R
\name{generate_synthetic_data}
\alias{generate_synthetic_data}
\title{Generate synthetic GreenFeed data for testing or simulation.}
\usage{
generate_synthetic_data(path, n, seed)
}
\arguments{
\item{path}{A path to where the synthetic data should be saved.}

\item{n}{The number of rows of synthetic data to generate.}

\item{seed}{An optional seed for reproducibility.}
}
\value{
A CSV file containing synthetic GreenFeed data.
The data includes the following columns:
\itemize{
\item FeederID: A numeric value corresponding to the feeder ID.
\item AnimalName: A character value corresponding to the animal name.
\item RFID: A character value corresponding to the RFID.
\item StartTime: A datetime value corresponding to the start time.
\item EndTime: A datetime value corresponding to the end time.
\item GoodDataDuration: A time value corresponding to the good data duration.
\item CO2GramsPerDay: A numeric value corresponding to the CO2 grams per day.
\item CH4GramsPerDay: A numeric value corresponding to the CH4 grams per day.
\item O2GramsPerDay: A numeric value corresponding to the O2 grams per day.
\item H2GramsPerDay: A numeric value corresponding to the H2 grams per day.
\item H2SGramsPerDay: A numeric value corresponding to the H2S grams per day.
\item AirflowLitersPerSec: A numeric value corresponding to the airflow in liters per second.
\item AirflowCf: A numeric value corresponding to the airflow in cubic feet.
\item WindSpeedMetersPerSec: A numeric value corresponding to the wind speed in meters per second.
\item WindDirDeg: A numeric value corresponding to the wind direction in degrees.
\item WindCf: A numeric value corresponding to the wind in cubic feet.
\item WasInterrupted: A logical value indicating whether the data was interrupted.
\item InterruptingTags: A character value corresponding to the interrupting tags.
\item TempPipeDegreesCelsius: A numeric value corresponding to the temperature of the pipe in degrees Celsius.
\item IsPreliminary: A logical value indicating whether the data is preliminary.
\item RunTime: A datetime value corresponding to the run time.
}
}
\description{
Generate synthetic GreenFeed data for testing or simulation.
}
\examples{
\dontrun{
generate_synthetic_data("data/synthetic_prelim_greenfeed_data.csv", 1000)
 }
}
