% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_validation.R
\name{get_eid_type}
\alias{get_eid_type}
\title{Get the Manufacturer or Type of an EID}
\usage{
get_eid_type(eid)
}
\arguments{
\item{eid}{A character string representing
the electronic identification number (EID) of an animal.}
}
\value{
A character string representing the manufacturer or type of the EID.
}
\description{
Get the Manufacturer or Type of an EID
}
\examples{
get_eid_type("982012345678901") # returns "Allflex"
get_eid_type("985012345678901") # returns "Destron Fearing"
get_eid_type("840012345678901") # returns "USDA Animal Identification Number"
get_eid_type("123456789012345") # returns NA_character_
get_eid_type(NA) # returns NA_character_
}
