% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/column_specs.R
\name{get_greenfeed_schema}
\alias{get_greenfeed_schema}
\title{Read in a CSV file containing the GreenFeed data schema.}
\usage{
get_greenfeed_schema(type = "preliminary")
}
\arguments{
\item{type}{Specifies the type of the greenfeed data to write a
column specification for. Options are \code{preliminary} and \code{verified}.}
}
\value{
A tibble with the standardized column names
for greenfeed data that has been read
in and parsed according to the
preliminary greenfeed data column specification.
}
\description{
Read in a CSV file containing the GreenFeed data schema.
}
\examples{
\dontrun{
get_greenfeed_schema()
}

}
