% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_col_spec.R
\name{get_prelim_data_schema}
\alias{get_prelim_data_schema}
\title{Get the preliminary greenfeed data schema from a CSV file}
\usage{
get_prelim_data_schema(none)
}
\arguments{
\item{none}{A placeholder for the function to work with the
get_prelim_data_schema function in the roxygen2 documentation.}
}
\value{
A tibble with the standardized column names
for preliminary greenfeed data that has been read
in and parsed according to the
preliminary greenfeed data column specification.
}
\description{
Get the preliminary greenfeed data schema from a CSV file
}
\examples{
\dontrun{
get_prelim_data_schema()
}

}
