% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grz_movement.R
\name{grz_movement_metrics}
\alias{grz_movement_metrics}
\title{Movement metrics by animal and period}
\usage{
grz_movement_metrics(data, by_period = c("total", "week", "date"))
}
\arguments{
\item{data}{Data frame of GPS fixes.}

\item{by_period}{Period grouping: \code{"total"}, \code{"week"}, \code{"date"}, or a vector
of these values.}
}
\value{
A \code{data.table} of movement metrics by entity and period.
}
\description{
Movement metrics by animal and period
}
