% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grz_qc.R
\name{grz_qc_sampling}
\alias{grz_qc_sampling}
\title{Sampling interval QC metrics}
\usage{
grz_qc_sampling(data, target_mins, tol_prop = 0.2)
}
\arguments{
\item{data}{Data frame of GPS fixes.}

\item{target_mins}{Target sampling interval in minutes.}

\item{tol_prop}{Proportional tolerance around target interval.}
}
\value{
A \code{data.table} with sampling QC metrics per entity.
}
\description{
Sampling interval QC metrics
}
