% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename_record_column.R
\name{rename_record_column}
\alias{rename_record_column}
\title{Rename "Record" Column}
\usage{
rename_record_column(df)
}
\arguments{
\item{df}{A data frame with exactly 1 column named "Record".}
}
\value{
A data frame.
}
\description{
Rename "Record" Column
}
\examples{
df <- data.frame(Record = 1)
rename_record_column(df)

}
