% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_ancestry_column.R
\name{split_ancestry_column}
\alias{split_ancestry_column}
\title{Split Ancestry Column into Allotment and Pasture}
\usage{
split_ancestry_column(df)
}
\arguments{
\item{df}{A data frame with exactly 1 column named "Ancestry".}
}
\description{
Split Ancestry Column into Allotment and Pasture
}
