% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strict_read_csv.R
\name{strict_read_csv}
\alias{strict_read_csv}
\title{strict_read_csv}
\usage{
strict_read_csv(data)
}
\arguments{
\item{data}{Path to a CSV file.}
}
\value{
A tibble with the data from the CSV file
that has been read in and parsed according to the
preliminary greenfeed data column specification.
}
\description{
strict_read_csv
}
\examples{
\dontrun{
strict_read_csv(here::here("data", "preliminary-greenfeed-data.csv"))
}
}
