% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/column_specs.R
\name{write_greenfeed_col_spec}
\alias{write_greenfeed_col_spec}
\title{Write GreenFeed column specification to an RDS file.}
\usage{
write_greenfeed_col_spec(file, type = "preliminary")
}
\arguments{
\item{file}{The name of the RDS file to write the column specification to.}

\item{type}{Specifies the type of the greenfeed data to write a
column specification for. Options are \code{preliminary} and \code{verified}.}
}
\value{
An RDS file containing the column specification for the
preliminary greenfeed data.
}
\description{
Write GreenFeed column specification to an RDS file.
}
\examples{
\dontrun{
write_greenfeed_col_spec("preliminary-greenfeed-col-spec.rds")
}
}
