% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_utilities.R
\name{compare_file_hashes}
\alias{compare_file_hashes}
\title{compare_file_hashes}
\usage{
compare_file_hashes(old_file, new_file)
}
\arguments{
\item{old_file}{Path to the old file.}

\item{new_file}{Path to the new file.}
}
\description{
compare_file_hashes
}
