% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_col_spec.R
\name{greenfeed_col_spec}
\alias{greenfeed_col_spec}
\title{Create a \code{readr} column specification for the preliminary greenfeed data}
\usage{
greenfeed_col_spec(none)
}
\arguments{
\item{none}{A placeholder for the function to work with the
create_col_spec function in the roxygen2 documentation.}
}
\value{
A \code{readr} column specification for the preliminary greenfeed data.
}
\description{
Create a \code{readr} column specification for the preliminary greenfeed data
}
\examples{
\dontrun{
create_col_spec()
}
}
