% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate-synthetic-data.R
\name{generate_eid}
\alias{generate_eid}
\title{Generate Electronic Identification (EID) numbers.}
\usage{
generate_eid(n)
}
\arguments{
\item{n}{A numeric value corresponding to the number of
EID (Electronic Identifier) numbers to generate.}
}
\value{
A vector of EID numbers. Each EID number is a 15-digit integer.
The EID numbers are generated randomly and are unique within the set.
}
\description{
Generate Electronic Identification (EID) numbers.
}
\examples{
generate_eid(n = 10)

}
