% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forage_production.R
\name{generate_forage_data}
\alias{generate_forage_data}
\title{Generate a Random Sample of Forage Production Data}
\usage{
generate_forage_data(n_rows, mean = 100, sd = 50)
}
\arguments{
\item{n_rows}{The number of samples (rows) to generate.}

\item{mean}{The mean value for the forage samples.}

\item{sd}{The standard deviation for the forage samples.}
}
\value{
A data frame with simulated forage production data.
The data frame will have columns named Sample_1, Sample_2, ..., Sample_5.
Each 'Sample_' column represents a dried forage sample in grams.
}
\description{
Generate a Random Sample of Forage Production Data
}
\examples{
generate_forage_data(n_rows = 10)

}
