% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_utilities.R
\name{generate_sha256_hash}
\alias{generate_sha256_hash}
\title{Generate sha256 hash for a file}
\usage{
generate_sha256_hash(file_path)
}
\arguments{
\item{file_path}{A path to a file.}
}
\value{
A sha256 hash for the file.
}
\description{
Generate sha256 hash for a file
}
\examples{
temp_file <- tempfile(fileext = ".txt")
writeLines("Hello, world!", temp_file)
sha256_hash <- generate_sha256_hash(temp_file)

}
