% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_validation.R
\name{eid_length}
\alias{eid_length}
\title{Get Length of EID}
\usage{
eid_length(eid)
}
\arguments{
\item{eid}{A character string representing
the electronic identification number (EID) of an animal.}
}
\value{
An integer representing the length of the EID.
}
\description{
Get Length of EID
}
\examples{
eid_length("123456789012345") # returns 15

}
