% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_col_spec.R
\name{get_prelim_col_spec}
\alias{get_prelim_col_spec}
\title{Get the preliminary greenfeed data column specification from an RDS file}
\usage{
get_prelim_col_spec(none)
}
\arguments{
\item{none}{A placeholder for the function to work with the
get_prelim_data_schema function in the roxygen2 documentation.}
}
\value{
A readr column specification for the
preliminary greenfeed data where each column is parsed
according to the data type in the column specification.
}
\description{
Get the preliminary greenfeed data column specification from an RDS file
}
\examples{
\dontrun{
get_prelim_col_spec()
}

}
