% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_validation.R
\name{validate_eid_length}
\alias{validate_eid_length}
\title{Validate EID Length}
\usage{
validate_eid_length(eid)
}
\arguments{
\item{eid}{A character string representing the electronic identification number (EID) of an animal.}
}
\value{
A logical value indicating whether the EID is of the correct length.
}
\description{
Validate EID Length
}
