% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_validation.R
\name{validate_eid_prefix}
\alias{validate_eid_prefix}
\title{Validate EID Prefix}
\usage{
validate_eid_prefix(eid)
}
\arguments{
\item{eid}{A character string representing the electronic identification number (EID) of an animal.}
}
\value{
A logical value indicating whether the EID has the correct prefix.
}
\description{
Validate EID Prefix
}
